Implementar AtencionService y AtencionController con logica de turnos

+++ src/main/java/com/consultorio/backend/repositories/PacienteRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.consultorio.backend.entities.Paciente;
 
-public interface PacienteRepository extends JpaRepository<Paciente, Integer> {
+public interface PacienteRepository extends JpaRepository<Paciente, Long> {
     // Aquí irán tus Query Methods personalizados cuando los necesitemos
 }
